feature

+++ src/main/java/ua/rstkhldntsk/servlet/services/InvoiceService.java
@@ -5,9 +5,9 @@ import ua.rstkhldntsk.servlet.dao.DaoFactory;
 import ua.rstkhldntsk.servlet.dao.interfaces.InvoiceDAO;
 import ua.rstkhldntsk.servlet.dao.interfaces.ProductDAO;
 import ua.rstkhldntsk.servlet.exceptions.IdNotExist;
-import ua.rstkhldntsk.servlet.exceptions.InvalidInput;
 import ua.rstkhldntsk.servlet.exceptions.ProductAlreadyExistException;
 import ua.rstkhldntsk.servlet.exceptions.NotEnoughProduct;
+import ua.rstkhldntsk.servlet.exceptions.ProductNotExist;
 import ua.rstkhldntsk.servlet.models.Invoice;
 import ua.rstkhldntsk.servlet.models.Product;
 import ua.rstkhldntsk.servlet.models.User;
@@ -51,11 +51,24 @@ public class InvoiceService {
         }
     }
 
-    public void deleteProductFromInvoice(String productName, Integer invoiceId, Integer langId) {
-        Product product = productDAO.findByName(productName, langId).get();
-        invoiceDAO.deleteProductFromInvoice(product.getCode(), invoiceId);
+    public void deleteProductFromInvoice(String productName, Integer invoiceId, Integer langId) throws ProductNotExist {
+        Optional<Product> productOptional = productDAO.findByName(productName, langId);
         Invoice invoice = invoiceDAO.findById(invoiceId, 1).get();
-        invoiceDAO.update(invoice);
+        if (productOptional.isPresent()) {
+            Product product = productOptional.get();
+            if ((long) invoice.getProducts().keySet().size() > 1) {
+                invoiceDAO.deleteProductFromInvoice(product.getCode(), invoiceId);
+                invoiceDAO.update(invoice);
+                LOGGER.debug(product.getName() + " deleted");
+                return;
+            }
+            invoiceDAO.deleteProductFromInvoice(product.getCode(), invoiceId);
+            invoiceDAO.delete(invoice);
+            LOGGER.debug("Product " + product.getCode() + " and invoice " + invoice.getId() + " were deleted");
+        } else {
+            LOGGER.error("This product isn't in this invoice");
+            throw new ProductNotExist();
+        }
     }
 
     public void updateInvoice(Invoice invoice) {
